Details in html files
The starting points of an HTML file has been added to all the files and the starting point of our main page is coded and ready to use JSON files to load content in other files as well as index.html.

From now on we can start working on our code and get into the working part of our code.

Signed-off-by: Arad190783 <noar24xe@student.ju.se>